worked on World Rotation
By pressing Q and E the world can rotate now around the position of the player. Before starting the rotation the physics calculations are stopped for the entire world.
After finishing the rotation the physics calculations are resumed. Momentum is preserved for all objects.
But the movement of the player cannot be done by just manipulating the velocity directly. It needs to manipulate the forces, otherwise it won't preserve momentum. (see moveTest Script)